Renamed RegattaStore.dart to naming convention.
