+++ frontend/lib/app_component.dart
@@ -5,7 +5,7 @@ import 'package:greencat/greencat.dart';
 
 import 'event_list_component.dart';
 import 'event_detail_component.dart';
-import 'package:frontend/store/RegattaStore.dart';
+import 'package:frontend/store/regatta_store.dart';
 import 'package:frontend/services/event_service.dart';
 import 'package:frontend/services/sailing_club_service.dart';
 import 'package:frontend/services/person_service.dart';
