+++ frontend/lib/store/regatta_store.dart
